backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pypdf==3.17.1
chromadb==0.4.18
sentence-transformers==2.2.2
openai==1.3.7
numpy==1.24.3
python-dotenv==1.0.0
python-multipart==0.0.6
torch==2.1.0